+++ extensions/apps/extensions/src/components/index.tsx
@@ -1,29 +1,42 @@
 import React from 'react';
-import ReactDOMClient from 'react-dom/client';
-import singleSpaReact from 'single-spa-react';
+import { I18nextProvider } from 'react-i18next';
+import { useRootComponentProps, withProviders } from '@akashaorg/ui-core-hooks';
+import Spinner from '@akashaorg/design-system-core/lib/components/Spinner';
+import { RouterProvider } from '@tanstack/react-router';
+import { router } from './app-routes/index';
+import { useApolloClient } from '@apollo/client';
+import { Helmet, helmetData } from '@akashaorg/design-system-core/lib/utils';
 
-import { IRootComponentProps } from '@akashaorg/typings/lib/ui';
-import ErrorLoader from '@akashaorg/design-system-core/lib/components/ErrorLoader';
-import { withProviders } from '@akashaorg/ui-core-hooks';
+declare module '@tanstack/react-router' {
+  interface Register {
+    router: ReturnType<typeof router>;
+  }
+}
 
-import App from './app';
+const App: React.FC<unknown> = () => {
+  const { getTranslationPlugin, baseRouteName, worldConfig, decodeAppName, plugins } =
+    useRootComponentProps();
+  const apolloClient = useApolloClient();
 
-const reactLifecycles = singleSpaReact({
-  React,
-  ReactDOMClient,
-  rootComponent: withProviders(App),
-  errorBoundary: (error, errorInfo, props: IRootComponentProps) => {
-    if (props.logger) {
-      props.logger.error(`${JSON.stringify(error)}, ${errorInfo}`);
-    }
-    return (
-      <ErrorLoader type="script-error" title="Error in extensions app" details={error.message} />
-    );
-  },
-});
+  return (
+    <React.StrictMode>
+      <React.Suspense fallback={<Spinner />}>
+        <I18nextProvider i18n={getTranslationPlugin().i18n}>
+          <Helmet helmetData={helmetData}>
+            <title>Extensions | {worldConfig.title}</title>
+          </Helmet>
+          <RouterProvider
+            router={router({
+              baseRouteName,
+              apolloClient,
+              decodeAppName,
+              plugins,
+            })}
+          />
+        </I18nextProvider>
+      </React.Suspense>
+    </React.StrictMode>
+  );
+};
 
-export const bootstrap = reactLifecycles.bootstrap;
-
-export const mount = reactLifecycles.mount;
-
-export const unmount = reactLifecycles.unmount;
+export default withProviders(App);
