--- extensions/apps/legal/src/components/App.tsx
@@ -1,30 +1,0 @@
-import * as React from 'react';
-import { I18nextProvider } from 'react-i18next';
-import { useRootComponentProps } from '@akashaorg/ui-core-hooks';
-import Spinner from '@akashaorg/design-system-core/lib/components/Spinner';
-import { RouterProvider } from '@tanstack/react-router';
-import { router } from './app-routes';
-
-declare module '@tanstack/react-router' {
-  interface Register {
-    router: ReturnType<typeof router>;
-  }
-}
-
-const App: React.FC<unknown> = () => {
-  const { getTranslationPlugin, baseRouteName } = useRootComponentProps();
-
-  return (
-    <React.Suspense fallback={<Spinner />}>
-      <I18nextProvider i18n={getTranslationPlugin().i18n}>
-        <RouterProvider
-          router={router({
-            baseRouteName,
-          })}
-        />
-      </I18nextProvider>
-    </React.Suspense>
-  );
-};
-
-export default App;
