+++ extensions/apps/search/src/components/index.tsx
@@ -1,27 +1,34 @@
 import React from 'react';
-import ReactDOMClient from 'react-dom/client';
-import singleSpaReact from 'single-spa-react';
+import { I18nextProvider } from 'react-i18next';
+import { useRootComponentProps, withProviders } from '@akashaorg/ui-core-hooks';
+import Spinner from '@akashaorg/design-system-core/lib/components/Spinner';
+import Stack from '@akashaorg/design-system-core/lib/components/Stack';
+import { RouterProvider } from '@tanstack/react-router';
+import { router } from './app-routes';
+import { useApolloClient } from '@apollo/client';
+import { Helmet, helmetData } from '@akashaorg/design-system-core/lib/utils';
 
-import { IRootComponentProps } from '@akashaorg/typings/lib/ui';
-import ErrorLoader from '@akashaorg/design-system-core/lib/components/ErrorLoader';
-import { withProviders } from '@akashaorg/ui-core-hooks';
+const App: React.FC<unknown> = () => {
+  const { getTranslationPlugin, baseRouteName, worldConfig } = useRootComponentProps();
+  const apolloClient = useApolloClient();
 
-import App from './App';
+  return (
+    <React.Suspense fallback={<Spinner />}>
+      <I18nextProvider i18n={getTranslationPlugin().i18n}>
+        <Helmet helmetData={helmetData}>
+          <title>Search | {worldConfig.title}</title>
+        </Helmet>
+        <Stack dataTestId="search-box">
+          <RouterProvider
+            router={router({
+              baseRouteName,
+              apolloClient,
+            })}
+          />
+        </Stack>
+      </I18nextProvider>
+    </React.Suspense>
+  );
+};
 
-const reactLifecycles = singleSpaReact({
-  React,
-  ReactDOMClient,
-  rootComponent: withProviders(App),
-  errorBoundary: (error, errorInfo, props: IRootComponentProps) => {
-    if (props.logger) {
-      props.logger.error(`${JSON.stringify(error)}, ${errorInfo}`);
-    }
-    return <ErrorLoader type="script-error" title="Error in search app" details={error.message} />;
-  },
-});
-
-export const bootstrap = reactLifecycles.bootstrap;
-
-export const mount = reactLifecycles.mount;
-
-export const unmount = reactLifecycles.unmount;
+export default withProviders(App);
