+++ extensions/apps/settings/src/components/index.tsx
@@ -1,29 +1,37 @@
-import React from 'react';
-import ReactDOMClient from 'react-dom/client';
-import singleSpaReact from 'single-spa-react';
+import * as React from 'react';
+import { I18nextProvider } from 'react-i18next';
+import { useRootComponentProps, withProviders } from '@akashaorg/ui-core-hooks';
+import Spinner from '@akashaorg/design-system-core/lib/components/Spinner';
 
-import { IRootComponentProps } from '@akashaorg/typings/lib/ui';
-import ErrorLoader from '@akashaorg/design-system-core/lib/components/ErrorLoader';
-import { withProviders } from '@akashaorg/ui-core-hooks';
+import { RouterProvider } from '@tanstack/react-router';
+import { router } from './app-routes';
+import { Helmet, helmetData } from '@akashaorg/design-system-core/lib/utils';
 
-import App from './app';
+declare module '@tanstack/react-router' {
+  interface Register {
+    router: ReturnType<typeof router>;
+  }
+}
 
-const reactLifecycles = singleSpaReact({
-  React,
-  ReactDOMClient,
-  rootComponent: withProviders(App),
-  errorBoundary: (error, errorInfo, props: IRootComponentProps) => {
-    if (props.logger) {
-      props.logger.error(`${JSON.stringify(error)}, ${errorInfo}`);
-    }
-    return (
-      <ErrorLoader type="script-error" title="Error in settings app" details={error.message} />
-    );
-  },
-});
+const Application: React.FC<unknown> = () => {
+  const { getTranslationPlugin, baseRouteName, worldConfig } = useRootComponentProps();
 
-export const bootstrap = reactLifecycles.bootstrap;
+  return (
+    <React.StrictMode>
+      <React.Suspense fallback={<Spinner />}>
+        <I18nextProvider i18n={getTranslationPlugin().i18n}>
+          <Helmet helmetData={helmetData}>
+            <title>Settings | {worldConfig.title}</title>
+          </Helmet>
+          <RouterProvider
+            router={router({
+              baseRouteName,
+            })}
+          />
+        </I18nextProvider>
+      </React.Suspense>
+    </React.StrictMode>
+  );
+};
 
-export const mount = reactLifecycles.mount;
-
-export const unmount = reactLifecycles.unmount;
+export default withProviders(Application);
