+++ extensions/widgets/sidebar/src/index.tsx
@@ -3,5 +3,5 @@ import { IntegrationRegistrationOptions, WidgetInterface } from '@akashaorg/typi
 
 export const register = (opts: IntegrationRegistrationOptions): WidgetInterface => ({
   mountsIn: opts.layoutSlots?.sidebarSlotId,
-  loadingFn: () => import('./sidebar'),
+  rootComponent: () => import('./sidebar'),
 });
