+++ extensions/widgets/sidebar/src/sidebar.tsx
@@ -1,9 +1,4 @@
 import React from 'react';
-import ReactDOMClient from 'react-dom/client';
-import singleSpaReact from 'single-spa-react';
-
-import { IRootComponentProps } from '@akashaorg/typings/lib/ui';
-import ErrorLoader from '@akashaorg/design-system-core/lib/components/ErrorLoader';
 import { useRootComponentProps, withProviders } from '@akashaorg/ui-core-hooks';
 import { I18nextProvider } from 'react-i18next';
 import SidebarComponent from './components/sidebar-component';
@@ -17,17 +12,4 @@ const Widget: React.FC<unknown> = () => {
     </I18nextProvider>
   );
 };
-
-export const { bootstrap, mount, unmount } = singleSpaReact({
-  React,
-  ReactDOMClient,
-  rootComponent: withProviders(Widget),
-  errorBoundary: (error, errorInfo, props: IRootComponentProps) => {
-    if (props.logger) {
-      props.logger.error(`${JSON.stringify(error)}, ${errorInfo}`);
-    }
-    return (
-      <ErrorLoader type="script-error" title="Error in sidebar widget" details={error.message} />
-    );
-  },
-});
+export default withProviders(Widget);
