+++ extensions/widgets/top-bar/src/components/index.tsx
@@ -1,28 +1,23 @@
 import React from 'react';
-import ReactDOMClient from 'react-dom/client';
-import singleSpaReact from 'single-spa-react';
-import { IRootComponentProps } from '@akashaorg/typings/lib/ui';
-import ErrorLoader from '@akashaorg/design-system-core/lib/components/ErrorLoader';
-import { withProviders } from '@akashaorg/ui-core-hooks';
+import { I18nextProvider } from 'react-i18next';
+import { RouterProvider } from '@tanstack/react-router';
+import { useRootComponentProps, withProviders } from '@akashaorg/ui-core-hooks';
+import { router } from './widget-routes';
 
-import Widget from './topbar-widget';
+declare module '@tanstack/react-router' {
+  interface Register {
+    router: typeof router;
+  }
+}
 
-const reactLifecycles = singleSpaReact<IRootComponentProps>({
-  React,
-  ReactDOMClient,
-  rootComponent: withProviders(Widget),
-  errorBoundary: (error, errorInfo, props: IRootComponentProps) => {
-    if (props.logger) {
-      props.logger.error(`${JSON.stringify(error)}, ${errorInfo}`);
-    }
-    return (
-      <ErrorLoader type="script-error" title="Error in topbar widget" details={error.message} />
-    );
-  },
-});
+const TopbarWidget: React.FC<unknown> = () => {
+  const { getTranslationPlugin } = useRootComponentProps();
 
-export const bootstrap = reactLifecycles.bootstrap;
+  return (
+    <I18nextProvider i18n={getTranslationPlugin().i18n}>
+      <RouterProvider router={router} />
+    </I18nextProvider>
+  );
+};
 
-export const mount = reactLifecycles.mount;
-
-export const unmount = reactLifecycles.unmount;
+export default withProviders(TopbarWidget);
