+++ extensions/widgets/trending/src/components/index.tsx
@@ -1,28 +1,20 @@
 import React from 'react';
-import ReactDOMClient from 'react-dom/client';
-import singleSpaReact from 'single-spa-react';
-import { IRootComponentProps } from '@akashaorg/typings/lib/ui';
-import ErrorLoader from '@akashaorg/design-system-core/lib/components/ErrorLoader';
-import { withProviders } from '@akashaorg/ui-core-hooks';
+import { I18nextProvider } from 'react-i18next';
 
-import App from './app';
+import { useRootComponentProps, withProviders } from '@akashaorg/ui-core-hooks';
 
-const reactLifecycles = singleSpaReact({
-  React,
-  ReactDOMClient,
-  rootComponent: withProviders(App),
-  errorBoundary: (error, errorInfo, props: IRootComponentProps) => {
-    if (props.logger) {
-      props.logger.error(`${JSON.stringify(error)}, ${errorInfo}`);
-    }
-    return (
-      <ErrorLoader type="script-error" title="Error in trending widget" details={error.message} />
-    );
-  },
-});
+import TrendingWidgetComponent from './trending-widget-component';
 
-export const bootstrap = reactLifecycles.bootstrap;
+const TrendingWidgetRoot: React.FC<unknown> = () => {
+  const { getTranslationPlugin } = useRootComponentProps();
 
-export const mount = reactLifecycles.mount;
+  return (
+    <>
+      <I18nextProvider i18n={getTranslationPlugin().i18n} defaultNS="ui-widget-trending">
+        <TrendingWidgetComponent />
+      </I18nextProvider>
+    </>
+  );
+};
 
-export const unmount = reactLifecycles.unmount;
+export default withProviders(TrendingWidgetRoot);
