+++ extensions/widgets/trending/src/index.tsx
@@ -2,6 +2,6 @@ import 'systemjs-webpack-interop/auto-public-path';
 import { IntegrationRegistrationOptions, WidgetInterface } from '@akashaorg/typings/lib/ui';
 
 export const register = (opts: IntegrationRegistrationOptions): WidgetInterface => ({
-  loadingFn: () => import('./components'),
+  rootComponent: () => import('./components'),
   mountsIn: opts.layoutSlots?.widgetSlotId,
 });
